Separate Vue Components
This patch moves several components to their own component files instead
of keeping everything in one huge JavaScript file.
